core update
commit1:
https://github.com/TrinityCore/TrinityCore/commit/62adbdab992ec177f60dec76447508741de4e7db#
比照TC過去提基修復來優化核心

commit2:
1.參考TC兒童週
https://github.com/TrinityCore/TrinityCore/commit/9c4cf9b39c17854fac3c1323e5e3bd1d2f164e3d#diff-f7bd8139463608ec43535300fae1ee00bde1ee756f901784729156edf8bb2fed
SCRIPT加入完成
SQL部分還沒處理
2.加入安其拉神廟AQ20無疤者無疤者奧斯里安腳本
代辦事項：不確定DB是否要加資料，尚需確認

+++ src/server/game/World/World.h
@@ -1,10 +1,21 @@
-////////////////////////////////////////////////////////////////////////////////
-//
-//  MILLENIUM-STUDIO
-//  Copyright 2016 Millenium-studio SARL
-//  All Rights Reserved.
-//
-////////////////////////////////////////////////////////////////////////////////
+/*
+* Copyright (C) 2008-2020 TrinityCore <http://www.trinitycore.org/>
+* Copyright (C) 2005-2009 MaNGOS <http://getmangos.com/>
+* Copyright (C) 2021 WodCore Reforged
+*
+* This program is free software; you can redistribute it and/or modify it
+* under the terms of the GNU General Public License as published by the
+* Free Software Foundation; either version 2 of the License, or (at your
+* option) any later version.
+*
+* This program is distributed in the hope that it will be useful, but WITHOUT
+* ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or
+* FITNESS FOR A PARTICULAR PURPOSE. See the GNU General Public License for
+* more details.
+*
+* You should have received a copy of the GNU General Public License along
+* with this program. If not, see <http://www.gnu.org/licenses/>.
+*/
 
 /// \addtogroup world The World
 /// @{
@@ -850,7 +861,7 @@ class World
         /// Get the maximum skill level a player can reach
         uint16 GetConfigMaxSkillValue() const
         {
-            uint8 lvl = uint8(getIntConfig(CONFIG_MAX_PLAYER_LEVEL));
+            uint16 lvl = uint16(getIntConfig(CONFIG_MAX_PLAYER_LEVEL));
 
             if (lvl >= 1 && lvl < 10)
                 return 75;
